feat(payment-providers): Consolidate Flutterwave service functionality into a single class and remove the old implementation
